Add _check_picks_uniqueness for extract_gfp_peaks
